requirements/base.txt
wheel==0.35
Django==3.1
djangosaml2==0.50.0
django_crispy_forms==1.10.0
django-filter==2.4.0
django-tables2==2.3.3
django-userforeignkey==0.3.0